+++ MOP/src/MOP.cs
@@ -62,13 +62,13 @@ namespace MOP
         static internal SettingRadioButtons PerformanceModes, Resolution;
         static internal SettingToggle EnableShadowAdjusting, KeepRunningInBackground,
                                       DynamicDrawDistance, RulesAutoUpdate, VerifyRuleFiles, DeleteUnusedRules,
-                                      DestroyEmptyBottles, DisableEmptyItems;
+                                      DestroyEmptyBottles, DisableEmptyItems, FasterAlgo;
 #else
         static internal SettingsSliderInt ActiveDistance, FramerateLimiter, ShadowDistance, RulesAutoUpdateFrequency;
         static internal SettingsCheckBoxGroup ModePerformance, ModeBalanced, ModeQuality, ModeSafe;
         static internal SettingsCheckBox KeepRunningInBackground, LimitFramerate, DynamicDrawDistance,
                                           RulesAutoUpdate, VerifyRuleFiles, DeleteUnusedRules,
-                                          DestroyEmptyBottles, DisableEmptyItems;
+                                          DestroyEmptyBottles, DisableEmptyItems, FasterAlgo;
 #endif
 
         readonly string[] activeDistanceText = { "Close (0.75x)", "Normal (1x)", "Far (2x)", "Very Far (4x)" };
@@ -125,6 +125,10 @@ namespace MOP
                 "<color=yellow>BALANCED (recommended)</color>: <color=white>Maintains balance between PERFORMANCE and QUALITY</color>\n" +
                 "<color=yellow>QUALITY</color>: <color=white>Hides obvious on-screen spawning and despawning, at the cost of performance</color>\n" +
                 "<color=yellow>SAFE</color>: <color=white>Despawns only minimum number of objects that are known to not cause any issues</color>");
+            modSettings.AddSpacer(10);
+            FasterAlgo = modSettings.AddToggle("fastAlgo", "<color=yellow>BETA</color>: FAST ALGORITHM", false);
+            modSettings.AddText("FAST ALGORITHM is an experimental function that is supposed to decrease the time it takes for MOP to toggle objects on and off. " +
+                "It will decrease the delay of enabling/disabling objects but might reduce the framerate.");
 
             // Graphics
             modSettings.AddHeader("GRAPHICS");
@@ -210,6 +214,10 @@ namespace MOP
             ModeBalanced = Settings.AddCheckBoxGroup(this, "modeBalanced", "BALANCED", true, "performanceMode", MopSettings.UpdatePerformanceMode);
             ModeQuality = Settings.AddCheckBoxGroup(this, "modeQuality", "QUALITY", false, "performanceMode", MopSettings.UpdatePerformanceMode);
             ModeSafe = Settings.AddCheckBoxGroup(this, "modeSafe", "<color=red>SAFE</color>", false, "performanceMode", MopSettings.UpdatePerformanceMode);
+            Settings.AddDynamicText(this, "\nWARNING ZONE:");
+            FasterAlgo = Settings.AddCheckBox(this, "fastAlgo", "<color=yellow>BETA</color>: FAST ALGORITHM", false);
+            Settings.AddDynamicText(this, "FAST ALGORITHM is an experimental function that is supposed to decrease the time it takes for MOP to toggle objects on and off. " +
+                "It will decrease the delay of enabling/disabling objects but might reduce the framerate.");
 
             // Graphics
             Settings.AddHeader(this, "GRAPHICS");