+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ### Changes
 
 - Very minor changes in exception throwing
+- Mod Loader Pro: Adjusted the color of the DONATE button
 
 ## 3.5.1 (28.02.2022)